join select fixed in articulos model

+++ PNotebook/PNotebook/ArticuloView.cs
@@ -14,7 +14,7 @@ namespace PNotebook
 			AppendColumn ("categoria", new CellRendererText (), "text", 2);
 			AppendColumn ("precio", new CellRendererText (), "text", 3);
 
-			Model = new ListStore (typeof(long), typeof(string), typeof(string), typeof(long));
+			Model = new ListStore (typeof(long), typeof(string), typeof(string), typeof(double));
 
 			refreshContent ();
 
@@ -27,7 +27,7 @@ namespace PNotebook
 			List<object[]> articulos = Articulos.listAll ();
 
 			foreach (object[] articulo in articulos) {
-				((ListStore)Model).AppendValues (Convert.ToInt64(articulo [0]), articulo [1], articulo[2], Convert.ToInt64(articulo[3]));
+				((ListStore)Model).AppendValues (Convert.ToInt64(articulo [0]), articulo [1], articulo[2], Convert.ToDouble(articulo[3]));
 			}
 		}
 	}
